+++ tests/lightning_base/test_measurements_class.py
@@ -27,7 +27,6 @@ from conftest import (  # tested device
     LightningStateVector,
     device_name,
 )
-from flaky import flaky
 from pennylane.devices import DefaultQubit
 from pennylane.measurements import VarianceMP
 from scipy.sparse import csr_matrix, random_array
@@ -619,7 +618,7 @@ class TestMeasurements:
         m = LightningMeasurements(statevector)
         return measure_final_state(m, tape)
 
-    @flaky(max_runs=5)
+    @pytest.mark.local_salt(42)
     @pytest.mark.parametrize("shots", [None, 500_000, [500_000, 500_000]])
     @pytest.mark.parametrize("measurement", [qml.expval, qml.probs, qml.var])
     @pytest.mark.parametrize(
@@ -642,7 +641,7 @@ class TestMeasurements:
             ),
         ),
     )
-    def test_single_return_value(self, shots, measurement, observable, lightning_sv, tol):
+    def test_single_return_value(self, shots, measurement, observable, lightning_sv, tol, seed):
         if obs_not_supported_in_ltensor(observable):
             pytest.skip("Observable not supported in lightning.tensor.")
 
@@ -671,7 +670,7 @@ class TestMeasurements:
 
         n_qubits = 4
         n_layers = 1
-        np.random.seed(0)
+        np.random.seed(seed)
         weights = np.random.rand(n_layers, n_qubits, 3)
         ops = [qml.Hadamard(i) for i in range(n_qubits)]
         if device_name != "lightning.tensor":
@@ -683,7 +682,7 @@ class TestMeasurements:
         )
         tape = qml.tape.QuantumScript(ops, measurements, shots=shots)
 
-        statevector = lightning_sv(n_qubits)
+        statevector = lightning_sv(n_qubits, seed=seed)
         statevector = get_final_state(statevector, tape)
         m = LightningMeasurements(statevector)
 
@@ -713,7 +712,7 @@ class TestMeasurements:
             # allclose -> absolute(a - b) <= (atol + rtol * absolute(b))
             assert np.allclose(result, expected, rtol=dtol, atol=dtol)
 
-    @flaky(max_runs=5)
+    @pytest.mark.local_salt(42)
     @pytest.mark.parametrize("shots", [None, 400_000, (400_000, 400_000)])
     @pytest.mark.parametrize("measurement", [qml.expval, qml.probs, qml.var])
     @pytest.mark.parametrize(
@@ -750,7 +749,7 @@ class TestMeasurements:
             ),
         ),
     )
-    def test_double_return_value(self, shots, measurement, obs0_, obs1_, lightning_sv, tol):
+    def test_double_return_value(self, shots, measurement, obs0_, obs1_, lightning_sv, tol, seed):
         if obs_not_supported_in_ltensor(obs0_) or obs_not_supported_in_ltensor(obs1_):
             pytest.skip("Observable not supported in lightning.tensor.")
 
@@ -784,7 +783,7 @@ class TestMeasurements:
         measurements = [measurement(op=obs0_), measurement(op=obs1_)]
         tape = qml.tape.QuantumScript(ops, measurements, shots=shots)
 
-        statevector = lightning_sv(n_qubits)
+        statevector = lightning_sv(n_qubits, seed=seed)
         statevector = get_final_state(statevector, tape)
         m = LightningMeasurements(statevector)